--- i-like-this-page-backend/src/main/java/com/minhojang/ilikethispagebackend/common/dto/LikeRequestDto.java
@@ -1,18 +1,0 @@
-package com.minhojang.ilikethispagebackend.common.dto;
-
-import lombok.Getter;
-import lombok.ToString;
-
-@ToString
-@Getter
-public class LikeRequestDto {
-
-	private final String url;
-	private final String uuid;
-
-	public LikeRequestDto(String uuid, String url) {
-		this.url = url;
-		this.uuid = uuid;
-	}
-
-}
