--- i-like-this-page-backend/src/main/java/com/minhojang/ilikethispagebackend/common/dto/LikeResponseDto.java
@@ -1,13 +1,0 @@
-package com.minhojang.ilikethispagebackend.common.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-
-@Builder
-@Getter
-public class LikeResponseDto {
-	private String url;
-	private String uuid;
-	private long likeCount;
-	private boolean likeStatus;
-}
