--- i-like-this-page-backend/src/main/java/com/minhojang/ilikethispagebackend/common/util/StringUtils.java
@@ -1,13 +1,0 @@
-package com.minhojang.ilikethispagebackend.common.util;
-
-public class StringUtils {
-
-	public static boolean isEmpty(String str) {
-		return str == null || str.length() == 0;
-	}
-
-	public static boolean isNotEmpty(String str) {
-		return !isEmpty(str);
-	}
-
-}
