--- i-like-this-page-backend/src/main/java/com/minhojang/ilikethispagebackend/exception/InvalidArgumentException.java
@@ -1,13 +1,0 @@
-package com.minhojang.ilikethispagebackend.exception;
-
-public class InvalidArgumentException extends RuntimeException {
-
-	public InvalidArgumentException(String message) {
-		super(message);
-	}
-
-	public InvalidArgumentException(String message, Throwable cause) {
-		super(message, cause);
-	}
-
-}
