--- i-like-this-page-backend/src/main/java/com/minhojang/ilikethispagebackend/exception/JsonException.java
@@ -1,13 +1,0 @@
-package com.minhojang.ilikethispagebackend.exception;
-
-public class JsonException extends RuntimeException {
-
-	public JsonException(String message) {
-		super(message);
-	}
-
-	public JsonException(String message, Throwable cause) {
-		super(message, cause);
-	}
-
-}
