--- i-like-this-page-backend/src/main/java/com/minhojang/ilikethispagebackend/exception/NotFoundException.java
@@ -1,13 +1,0 @@
-package com.minhojang.ilikethispagebackend.exception;
-
-public class NotFoundException extends RuntimeException {
-
-	public NotFoundException(String message) {
-		super(message);
-	}
-
-	public NotFoundException(String message, Throwable cause) {
-		super(message, cause);
-	}
-
-}
