--- i-like-this-page-backend/src/main/java/com/minhojang/ilikethispagebackend/exception/UnsupportedMethodException.java
@@ -1,13 +1,0 @@
-package com.minhojang.ilikethispagebackend.exception;
-
-public class UnsupportedMethodException extends RuntimeException {
-
-	public UnsupportedMethodException(String message) {
-		super(message);
-	}
-
-	public UnsupportedMethodException(String message, Throwable cause) {
-		super(message, cause);
-	}
-
-}
