+++ i-like-this-page-backend/src/main/java/link/iltp/common/dto/LikeRequestDto.java
@@ -1,0 +1,18 @@
+package link.iltp.common.dto;
+
+import lombok.Getter;
+import lombok.ToString;
+
+@ToString
+@Getter
+public class LikeRequestDto {
+
+	private final String url;
+	private final String uuid;
+
+	public LikeRequestDto(String uuid, String url) {
+		this.url = url;
+		this.uuid = uuid;
+	}
+
+}
