+++ i-like-this-page-backend/src/main/java/link/iltp/common/dto/LikeResponseDto.java
@@ -1,0 +1,13 @@
+package link.iltp.common.dto;
+
+import lombok.Builder;
+import lombok.Getter;
+
+@Builder
+@Getter
+public class LikeResponseDto {
+	private String url;
+	private String uuid;
+	private long likeCount;
+	private boolean likeStatus;
+}
