+++ i-like-this-page-backend/src/main/java/link/iltp/common/util/StringUtils.java
@@ -1,0 +1,13 @@
+package link.iltp.common.util;
+
+public class StringUtils {
+
+	public static boolean isEmpty(String str) {
+		return str == null || str.length() == 0;
+	}
+
+	public static boolean isNotEmpty(String str) {
+		return !isEmpty(str);
+	}
+
+}
