+++ i-like-this-page-backend/src/main/java/link/iltp/exception/InvalidArgumentException.java
@@ -1,0 +1,13 @@
+package link.iltp.exception;
+
+public class InvalidArgumentException extends RuntimeException {
+
+	public InvalidArgumentException(String message) {
+		super(message);
+	}
+
+	public InvalidArgumentException(String message, Throwable cause) {
+		super(message, cause);
+	}
+
+}
