+++ i-like-this-page-backend/src/main/java/link/iltp/exception/JsonException.java
@@ -1,0 +1,13 @@
+package link.iltp.exception;
+
+public class JsonException extends RuntimeException {
+
+	public JsonException(String message) {
+		super(message);
+	}
+
+	public JsonException(String message, Throwable cause) {
+		super(message, cause);
+	}
+
+}
