+++ i-like-this-page-backend/src/main/java/link/iltp/exception/NotFoundException.java
@@ -1,0 +1,13 @@
+package link.iltp.exception;
+
+public class NotFoundException extends RuntimeException {
+
+	public NotFoundException(String message) {
+		super(message);
+	}
+
+	public NotFoundException(String message, Throwable cause) {
+		super(message, cause);
+	}
+
+}
