+++ i-like-this-page-backend/src/main/java/link/iltp/exception/UnsupportedMethodException.java
@@ -1,0 +1,13 @@
+package link.iltp.exception;
+
+public class UnsupportedMethodException extends RuntimeException {
+
+	public UnsupportedMethodException(String message) {
+		super(message);
+	}
+
+	public UnsupportedMethodException(String message, Throwable cause) {
+		super(message, cause);
+	}
+
+}
